ChatApp/requirements.txt
streamlit
streamlit-chat
streamlit>=1.25.0
openai>=1.0.0
groq>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0